Configure transmission & external drives
Since I want to run transmission as the pi user, I have to overwrite
some stuff the transmission role does. This means that ansible will
report changes on each run: The role changes folder permissions to the
debian-transmission user, and I change them back to the pi user. This is
not ideal, but that configuration for the roles is store in
non-overwritable variables :(