Always restart and ensure transmission permissions
The ansible role we use for transmission, uses the default transmission
user (hardcoded), and thus changes the permissions of the folders and
config files for the transmission service. We need to make sure we
always overwrite it to use our configured user, or the process won't
start. It's not ideal (idempotence and all), but it works.